+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Changelog
     - Removed ``category_icon`` from ``NewsEntry`` and ``News``.
     - Added ``big_icon_url`` and ``small_icon_url`` to ``NewsCategory``
     - ``Leaderboard.page`` renamed to ``current_page``.
+    - ``Highscores.page`` renamed to ``current_page``.
+    - Removed ``owner``, ``world``, ``status`` and ``type`` from ``GuildHouse``.
 
 
 .. v5.6.0